requirements.txt
fastapi
uvicorn
SQLAlchemy
mysql-connector-python
pydantic
python-dotenv